更新修改

+++ server/controller/userController.js
@@ -177,8 +177,17 @@ exports.outerConnectAction = function(app){
                 userInfoDao.findOneUserInfo(conditions1,dbHelper,function(result2){
                     var imgUrl = result2.result.image;
                     userInfoDao.removeUserInfo(conditions1,dbHelper,function(result1){
-                        fs.unlinkSync('./public'+imgUrl);
-                        res.json(result1);
+                        if(result1.success == 1){
+                            if(imgUrl != ''){
+                                fs.unlinkSync('./public'+imgUrl);
+                            }
+                            var conditions3 = {"userId":id};
+                            fairyDao.removeFairy(conditions3,dbHelper,function(result3){
+                                res.json(result3);
+                            });
+                        }else{
+                            res.json(result1);
+                        } 
                     });
                 });  
             }else{
